chore: TotalCountViewModel 파일 분리
